Refactor TPCH connector to produce pages by default
